+++ examples/energy_district/example_pareto_front.py
@@ -1,4 +1,4 @@
-from q100opt.setup_model import ParetoFront
+from q100opt.scenario_tools import ParetoFront
 from matplotlib import pyplot as plt
 from q100opt.setup_model import load_csv_data
